Phase 1: extract table of contents into features/toc.js
First of the previously-blocked coupled features, now that core/state.js and
platform/ios-chrome.js exist. buildToc / toggleToc / scheduleTocActiveHeading
Update are exported; renderTocNavigation and updateTocActiveHeading are private.
Reads shared TOC + view-mode state from `state`, isIOSNative from platform, and
the sheet/sync helpers from platform/ios-chrome; looks up its DOM by id.

Pure refactor. main.js ~1,943 -> ~1,837 lines. Build + lint clean; 297 tests pass.
